start extracting roast view from start view
