+++ miniweb/src/style.css
@@ -30,6 +30,11 @@ body {
   min-height: 100vh;
 }
 
+control_cluster {
+	place-content: align;
+	background-color: #ff0000;
+}
+
 h1 {
   font-size: 3.2em;
   line-height: 1.1;
@@ -70,7 +75,7 @@ button {
   font-size: 1em;
   font-weight: 500;
   font-family: inherit;
-  background-color: #1a1a1a;
+  background-color: #121212;
   cursor: pointer;
   transition: border-color 0.25s;
 }
@@ -91,14 +96,14 @@ button:focus-visible {
     color: #747bff;
   }
   button {
-    background-color: #f9f9f9;
+    background-color: #d1d1d1;
   }
 }
 
 /* Style the slider */
 input[type="range"] {
   -webkit-appearance: none;
-  width: 90%;
+  width: 220px;
   height: 6px;
   /*background: #ddd;*/
   border-radius: 3px;
